requirements.txt
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
chromadb==0.4.22
openai==1.7.2
python-dotenv==1.0.0
tiktoken==0.5.2
